@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -46,7 +46,6 @@ flutter_icons:
   image_path_ios: "assets/launcher/icon.png"
   image_path_android: "assets/launcher/icon.png"
   transparent_image_button: ^0.1.1+1
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
