@@ -49,8 +49,7 @@ class _MenuScreenState extends State<MenuScreen> {
     if (madeInitSound == false) {
       Future.delayed(Duration.zero, () async {
         madeInitSound = true;
-        bool newArchievements = await DBApi.db.newArchievements();
-        audioPlayerService.menuSound(newArchievements);
+        audioPlayerService.menuSound();
       });
     }
     var lockScreen = audioPlayerService.lockScreen;
@@ -73,11 +72,12 @@ class _MenuScreenState extends State<MenuScreen> {
                 // X BUTTON
                 Positioned(
                     left: SizeUtil.getDoubleByDeviceHorizontal(40),
-                    bottom: SizeUtil.getDoubleByDeviceVertical(580),
+                    top: SizeUtil.getDoubleByDeviceVertical(40),
                     child: IgnorePointer(
                         ignoring: lockScreen,
                         child: Container(
-                          width: SizeUtil.getDoubleByDeviceHorizontal(Constant.xButtonSize),
+                          width: SizeUtil.getDoubleByDeviceHorizontal(
+                              Constant.xButtonSize),
                           child: ExitButton(
                             closeApp: true,
                           ),
@@ -87,7 +87,8 @@ class _MenuScreenState extends State<MenuScreen> {
                     bottom: Constant.playButtonDistanceBottom,
                     right: Constant.playButtonDistanceRight,
                     child: PlayButton(
-                      size: SizeUtil.getDoubleByDeviceVertical(Constant.playButtonSize),
+                      size: SizeUtil.getDoubleByDeviceVertical(
+                          Constant.playButtonSize),
                       pushedName: ExplanationScreen.routeTag,
                       active: (!lockScreen && !allArchieved),
                       animationDone: true,
@@ -98,15 +99,18 @@ class _MenuScreenState extends State<MenuScreen> {
                   left: SizeUtil.getDoubleByDeviceHorizontal(740),
                   child: Center(
                       child: AchievementButtonList(
-                          animalwidth: SizeUtil.getDoubleByDeviceHorizontal(Constant.achievementButtonSize),
-                          animalheight: SizeUtil.getDoubleByDeviceHorizontal((Constant.achievementButtonSize - 6)))),
+                          animalwidth: SizeUtil.getDoubleByDeviceHorizontal(
+                              Constant.achievementButtonSize),
+                          animalheight: SizeUtil.getDoubleByDeviceHorizontal(
+                              (Constant.achievementButtonSize - 6)))),
                 ),
                 // Stars
                 Positioned(
                     top: SizeUtil.getDoubleByDeviceVertical(240),
                     left: SizeUtil.getDoubleByDeviceHorizontal(555),
                     height: 350,
-                    width: SizeUtil.getDoubleByDeviceHorizontal(Constant.starGifSize),
+                    width: SizeUtil.getDoubleByDeviceHorizontal(
+                        Constant.starGifSize),
                     child: Container(
                         child: FlareActor(
                       "assets/animation/stars.flr",
@@ -116,19 +120,28 @@ class _MenuScreenState extends State<MenuScreen> {
                 Positioned(
                     bottom: SizeUtil.getDoubleByDeviceVertical(95),
                     right: SizeUtil.getDoubleByDeviceHorizontal(875),
-                    child: Witch(standartWitchText: true, rotate: false, talking: false, size: Constant.witchSize)),
+                    child: Witch(
+                        standartWitchText: true,
+                        rotate: false,
+                        talking: false,
+                        size: Constant.witchSize)),
                 // WITCH
                 witchTalking
                     ? Positioned(
                         bottom: SizeUtil.getDoubleByDeviceVertical(95),
                         right: SizeUtil.getDoubleByDeviceHorizontal(875),
-                        child: Witch(rotate: false, talking: true, size: Constant.witchSize))
+                        child: Witch(
+                            rotate: false,
+                            talking: true,
+                            size: Constant.witchSize))
                     : Container(),
                 // ANIMAL
                 Positioned(
                     left: SizeUtil.getDoubleByDeviceHorizontal(500),
                     top: SizeUtil.getDoubleByDeviceVertical(475),
-                    child: IgnorePointer(ignoring: lockScreen, child: SelectedAnimal(size: Constant.menuAnimalSize))),
+                    child: IgnorePointer(
+                        ignoring: lockScreen,
+                        child: SelectedAnimal(size: Constant.menuAnimalSize))),
                 // CHANGE ANIMAL BUTTON
                 Positioned(
                     left: SizeUtil.getDoubleByDeviceHorizontal(400),
@@ -139,10 +152,12 @@ class _MenuScreenState extends State<MenuScreen> {
                           child: RawMaterialButton(
                             child: DarkableImage(
                                 url: 'assets/pics/reverse_blue.png',
-                                width: SizeUtil.getDoubleByDeviceHorizontal(Constant.changeAnimalButtonSize),
+                                width: SizeUtil.getDoubleByDeviceHorizontal(
+                                    Constant.changeAnimalButtonSize),
                                 fit: BoxFit.fitWidth),
                             onPressed: () {
-                              Navigator.pushNamed(context, "animalSelectionScreenRoute");
+                              Navigator.pushNamed(
+                                  context, "animalSelectionScreenRoute");
                             },
                           ),
                         ))),
